requirements.txt
Django==4.2.5
Pillow==10.0.1
django-admin-sortable2==2.1.9
django-extensions==3.2.3
django-tinymce==3.6.1
environs==9.5.0
whitenoise==6.6.0
requests==2.31.0
